@@ -1,14 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Xerum.XFramework.Common" Version="7.0.10" />
-    <PackageReference Include="Xerum.XFramework.Domain" Version="7.0.11" />
+    <PackageReference Include="Xerum.XFramework.Common" Version="8.1.0" />
   </ItemGroup>
 
 </Project>
