@@ -1,88 +1,43 @@
 ﻿using Xerum.XFramework.Common.Dto;
 using Xerum.XFramework.Common.Enums;
-using XRFID.Demo.Common.Enumerations;
 
 namespace XRFID.Demo.Common.Dto;
 
-public class MovementDto : RestEntityDto
+public class MovementDto : BaseDto
 {
     public int Sequence { get; set; }
+    public WorkflowType Type { get; set; }
+    public MovementContentEnum ContentType { get; set; }
+    public string? Description { get; set; }
+    public DateTimeOffset Timestamp { get; set; }
+    public bool IsValid { get; set; } = false;
+    public bool UnexpectedItem { get; set; } = false;
+    public bool MissingItem { get; set; } = false;
+    public bool OverflowItem { get; set; } = false;
+    public bool IsActive { get; set; } = false;
+    public bool IsConsolidated { get; set; } = false;
+    public bool Sent { get; set; } = false;
 
-    public string Description { get; set; } = string.Empty;
+    public Guid? ReaderId { get; set; }
 
-    public DateTime Timestamp { get; set; } = DateTime.Now;
+    public Guid? PrinterId { get; set; }
 
-    public bool IsValid { get; set; }
+    public Guid? BOMId { get; set; }
 
-    public bool UnexpectedItem { get; set; }
+    #region Custom attributes
+    public string? Attribute1 { get; set; }
+    public string? Attribute2 { get; set; }
+    #endregion
 
-    public bool MissingItem { get; set; }
+    #region Data Relations
 
-    public bool OverflowItem { get; set; }
+    public ReaderDto? Reader { get; set; }
 
-    public bool IsActive { get; set; }
+    public PrinterDto? Printer { get; set; }
 
-    public bool IsConsolidated { get; set; }
+    public ICollection<MovementItemDto> MovementItems { get; set; } = [];
 
-    /// <summary>
-    /// init, processing, completed
-    /// </summary>
-    public MovementStatus WorkingStatus
-    {
-        get
-        {
-            if (IsActive && !IsConsolidated)
-            {
-                if (MovementItems.Any())
-                {
-                    return MovementStatus.Processing;
-                }
-                else
-                {
-                    return MovementStatus.Initialize;
-                }
-            }
-            else
-            {
-                return MovementStatus.Completed;
-            }
-        }
-    }
+    public ICollection<LoadingUnitDto> LoadingUnits { get; set; } = [];
 
-    public MovementStatus Status
-    {
-        get
-        {
-            if (IsValid)
-            {
-                return MovementStatus.Valid;
-            }
-            else if (UnexpectedItem)
-            {
-                return MovementStatus.Unexpected;
-            }
-            else if (MissingItem)
-            {
-                return MovementStatus.Missing;
-            }
-            else if (OverflowItem)
-            {
-                return MovementStatus.Overflow;
-            }
-            else
-            {
-                return MovementStatus.Error;
-            }
-        }
-    }
-
-    public Guid ReaderId { get; set; }
-
-    public List<MovementItemDto> MovementItems { get; set; } = new List<MovementItemDto>();
-
-    public Guid? OrderId { get; set; }
-
-    public string? OrderReference { get; set; }
-
-    public MovementDirection Direction { get; set; } = MovementDirection.In;
+    #endregion
 }
