@@ -1,8 +1,10 @@
-﻿using Xerum.XFramework.MassTransit;
+﻿using XRFID.Demo.Common.Dto;
 using XRFID.Demo.Modules.Mqtt.Payloads;
 
 namespace XRFID.Demo.Modules.Mqtt.Publishers;
 
-public interface IZebraMqttCommandPublisher : IRequestPublisher<RAWMQTTCommands>
+public interface IZebraMqttCommandPublisher
 {
+    Task Publish(RAWMQTTCommands request);
+    Task SoundBuzzer(ReaderDto reader, RAWMQTTCommands command, uint onMilliseconds = 0);
 }
