@@ -1,9 +1,13 @@
-﻿using Xerum.XFramework.MassTransit;
-using XRFID.Demo.Modules.Mqtt.Contracts;
+﻿using XRFID.Demo.Modules.Mqtt.Contracts;
 using XRFID.Demo.Modules.Mqtt.Events;
 
 namespace XRFID.Demo.Modules.Mqtt.Publishers;
 
-public interface IZebraMqttEventPublisher : IRequestPublisher<Heartbeat>, IRequestPublisher<ZebraGpiData>, IRequestPublisher<ZebraTagData>
+public interface IZebraMqttEventPublisher
 {
+    Task Publish(Heartbeat request);
+    Task Publish(ZebraGpiData request);
+    Task Publish(ZebraTagData request);
+    Task Publish(ZebraDirectionalityTagData request);
+    Task Publish(ZebraRawDirectionalityTagData request);
 }
