@@ -1,9 +1,9 @@
-﻿using Xerum.XFramework.MassTransit;
-using XRFID.Demo.Modules.Mqtt.Payloads;
+﻿using XRFID.Demo.Modules.Mqtt.Payloads;
 
 namespace XRFID.Demo.Modules.Mqtt.Publishers;
 
-public interface IZebraMqttPayloadPublisher : IRequestPublisher<GetVersionResponse>,
-    IRequestPublisher<GetNetworkResponse>
+public interface IZebraMqttPayloadPublisher
 {
+    Task Publish(GetVersionResponse request);
+    Task Publish(GetNetworkResponse request);
 }
