@@ -4,21 +4,16 @@ using XRFID.Demo.Modules.Mqtt.Payloads;
 namespace XRFID.Demo.Modules.Mqtt.Publishers;
 
 
-public class ZebraMqttPayloadPublisher : IZebraMqttPayloadPublisher
+public class ZebraMqttPayloadPublisher(IBusControl _busControl) : IZebraMqttPayloadPublisher
 {
-    private readonly IBusControl busControl;
-    public ZebraMqttPayloadPublisher(IBusControl busControl)
-    {
-        this.busControl = busControl;
-    }
 
     public async Task Publish(GetVersionResponse request)
     {
-        await busControl.Publish<GetVersionResponse>(request);
+        await _busControl.Publish<GetVersionResponse>(request);
     }
 
     public async Task Publish(GetNetworkResponse request)
     {
-        await busControl.Publish<GetNetworkResponse>(request);
+        await _busControl.Publish<GetNetworkResponse>(request);
     }
 }
